// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "13.0.5";
  };
  public: {
    Tables: {
// ...
      experiences: {
        Row: {
          company: string;
          created_at: string;
          description: string | null;
          employment_type: string | null;
          highlights: string[] | null;
          id: string;
          location: string | null;
          order_index: number | null;
          period: string;
          position: string;
          updated_at: string;
        };
        Insert: {
          company: string;
          created_at?: string;
          description?: string | null;
          employment_type?: string | null;
          highlights?: string[] | null;
          id?: string;
          location?: string | null;
          order_index?: number | null;
          period: string;
          position: string;
          updated_at?: string;
        };
        Update: {
          company?: string;
          created_at?: string;
          description?: string | null;
          employment_type?: string | null;
          highlights?: string[] | null;
          id?: string;
          location?: string | null;
          order_index?: number | null;
          period?: string;
          position?: string;
          updated_at?: string;
        };
        Relationships: [];
      };
// ...
    };
// ...
  };
};